+++ examples/Cdd_Adc/Cdd_Adc_Example_GlbSwTrig/F29P32x/Cdd_Adc_Example_GlbSwTrig_Config/output/src/Cdd_Adc_Cfg.c
@@ -1,4 +1,3 @@
-
 /*********************************************************************************************************************
  *  COPYRIGHT
  *  ------------------------------------------------------------------------------------------------------------------
@@ -67,7 +66,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 {
     .hwunitcfg =
     {
-                
         [0] =
         {
             .hwunit_id = (Cdd_Adc_HwUnitType)CDD_ADCA,
@@ -93,14 +91,13 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif
             .startgroupnum = (Cdd_Adc_GroupType)(0U),
             .lastgroupnum = (Cdd_Adc_GroupType)(1U),
-            .base_addr = (uint32)(ADCA_BASE_FRAME(0U)),            
+            .base_addr = (uint32)(ADCA_BASE_FRAME(0U)),
             .result_baseaddr = (uint32)( ADCARESULT_BASE + (CDD_ADC_RESULTBASEADDR_STEP*0U)),
             .analogrefsel = (uint16)ASYSCTL_ANAREFCTL_ANAREFABSEL,
             .analogrefvoltsel = (uint16)ASYSCTL_ANAREFCTL_ANAREFAB_2P5SEL,
-            .inltrimaddress = ((uint32 *)&McalLib_DeviceCalibrationData.AdcAInlTrim[0U]),
+            .inltrimaddress =((uint32 *)&McalLib_DeviceCalibrationData.AdcAInlTrim[0U]),
             .numadc_inltrim = ((uint8)6U)
         },
-        
         [1] =
         {
             .hwunit_id = (Cdd_Adc_HwUnitType)CDD_ADCB,
@@ -126,14 +123,13 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif
             .startgroupnum = (Cdd_Adc_GroupType)(2U),
             .lastgroupnum = (Cdd_Adc_GroupType)(3U),
-            .base_addr = (uint32)(ADCB_BASE_FRAME(0U)),            
+            .base_addr = (uint32)(ADCB_BASE_FRAME(0U)),
             .result_baseaddr = (uint32)( ADCARESULT_BASE + (CDD_ADC_RESULTBASEADDR_STEP*1U)),
             .analogrefsel = (uint16)ASYSCTL_ANAREFCTL_ANAREFABSEL,
             .analogrefvoltsel = (uint16)ASYSCTL_ANAREFCTL_ANAREFAB_2P5SEL,
-            .inltrimaddress = ((uint32 *)&McalLib_DeviceCalibrationData.AdcBInlTrim[0U]),
+            .inltrimaddress =((uint32 *)&McalLib_DeviceCalibrationData.AdcBInlTrim[0U]),
             .numadc_inltrim = ((uint8)6U)
         },
-        
         [2] =
         {
             .hwunit_id = (Cdd_Adc_HwUnitType)CDD_ADCC,
@@ -159,16 +155,16 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif
             .startgroupnum = (Cdd_Adc_GroupType)(4U),
             .lastgroupnum = (Cdd_Adc_GroupType)(5U),
-            .base_addr = (uint32)(ADCC_BASE_FRAME(0U)),            
+            .base_addr = (uint32)(ADCC_BASE_FRAME(0U)),
             .result_baseaddr = (uint32)( ADCARESULT_BASE + (CDD_ADC_RESULTBASEADDR_STEP*2U)),
             .analogrefsel = (uint16)ASYSCTL_ANAREFCTL_ANAREFCDESEL,
             .analogrefvoltsel = (uint16)ASYSCTL_ANAREFCTL_ANAREFCDE_2P5SEL,
-            .inltrimaddress = ((uint32 *)&McalLib_DeviceCalibrationData.AdcCInlTrim[0U]),
+            .inltrimaddress =((uint32 *)&McalLib_DeviceCalibrationData.AdcCInlTrim[0U]),
             .numadc_inltrim = ((uint8)3U)
-        }        
+        }
     },
     .groupcfg  =
-    {        
+    {
         [0] =
         {
             /* HwunitId: CDD_ADCA */
@@ -191,7 +187,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .lastsocnum = (uint8)(1U),
             .dma_mode = ((boolean)(0U))
         },
-        
         [1] =
         {
             /* HwunitId: CDD_ADCA */
@@ -214,7 +209,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .lastsocnum = (uint8)(2U),
             .dma_mode = ((boolean)(0U))
         },
-        
         [2] =
         {
             /* HwunitId: CDD_ADCB */
@@ -237,7 +231,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .lastsocnum = (uint8)(1U),
             .dma_mode = ((boolean)(0U))
         },
-        
         [3] =
         {
             /* HwunitId: CDD_ADCB */
@@ -260,7 +253,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .lastsocnum = (uint8)(2U),
             .dma_mode = ((boolean)(0U))
         },
-        
         [4] =
         {
             /* HwunitId: CDD_ADCC */
@@ -283,7 +275,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .lastsocnum = (uint8)(1U),
             .dma_mode = ((boolean)(0U))
         },
-        
         [5] =
         {
             /* HwunitId: CDD_ADCC */
@@ -305,42 +296,40 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .soc_mask =  (uint32)(4U),
             .lastsocnum = (uint8)(2U),
             .dma_mode = ((boolean)(0U))
-        }        
+        }
     },
     .glbtrigcfg =
-    {                
+    {
         [0] =
-        {            
+        {
             .hwunit_mask = (uint8)(7U),
             .soc_mask = (uint32)(7U),
             .group_mask = (uint32)(63U)
         },
-        
         [1] =
-        {            
+        {
             .hwunit_mask = (uint8)(7U),
             .soc_mask = (uint32)(3U),
             .group_mask = (uint32)(21U)
         },
-        
         [2] =
-        {            
+        {
             .hwunit_mask = (uint8)(7U),
             .soc_mask = (uint32)(4U),
             .group_mask = (uint32)(42U)
         }
-    },    
+    },
     .glbsw_baseaddr = ((uint32)ADCGLOBAL_BASE_FRAME(0U)),
     .ppbcfg =
-    {                
+    {
         [0] =  
         {
             /* HwunitId: CDD_ADCA */
             .ppb_id = (Cdd_Adc_PpbIdType)(0U),
             .soc_num = (uint8)(0U),
             .hwunitindex = (uint8)(0U),
-            .ppb_caloffset = (uint32)(100U),
-            .ppb_refoffset = (uint32)(100U),
+            .ppb_caloffset = (uint16)(100U),
+            .ppb_refoffset = (uint16)(100U),
             .ppbresconfig = (uint16)((0x0000U) | (((uint16)0U)<<8U) | (((uint16)0U)<<7U) | (((uint16)0U)<<6U) | (((uint16)0U)<<5U)),
             .triplimithi = (sint32)30000,
             .triplimitlow = (sint32)1000,
@@ -355,15 +344,14 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .tripfilctl = (uint16)(0x0000U),     
             .tripfilclk = (uint16)(0x0000U)
         },
-
         [1] =  
         {
             /* HwunitId: CDD_ADCA */
             .ppb_id = (Cdd_Adc_PpbIdType)(1U),
             .soc_num = (uint8)(0U),
             .hwunitindex = (uint8)(0U),
-            .ppb_caloffset = (uint32)(100U),
-            .ppb_refoffset = (uint32)(100U),
+            .ppb_caloffset = (uint16)(100U),
+            .ppb_refoffset = (uint16)(100U),
             .ppbresconfig = (uint16)((0x0000U) | (((uint16)0U)<<8U) | (((uint16)0U)<<7U) | (((uint16)1U)<<6U) | (((uint16)0U)<<5U)),
             .triplimithi = (sint32)0,
             .triplimitlow = (sint32)-3000,
@@ -380,7 +368,7 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
         }
     },
     .channelcfg =
-    {        
+    {
         [0] =
         {
             /* HwunitId:CDD_ADCA Group:0 */
@@ -391,7 +379,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(0U)
         },
-        
         [1] =
         {
             /* HwunitId:CDD_ADCA Group:0 */
@@ -402,7 +389,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(1U)
         },
-        
         [2] =
         {
             /* HwunitId:CDD_ADCA Group:1 */
@@ -413,7 +399,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(2U)
         },
-        
         [3] =
         {
             /* HwunitId:CDD_ADCB Group:0 */
@@ -424,7 +409,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(0U)
         },
-        
         [4] =
         {
             /* HwunitId:CDD_ADCB Group:0 */
@@ -435,7 +419,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(1U)
         },
-        
         [5] =
         {
             /* HwunitId:CDD_ADCB Group:1 */
@@ -446,7 +429,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(2U)
         },
-        
         [6] =
         {
             /* HwunitId:CDD_ADCC Group:0 */
@@ -457,7 +439,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(0U)
         },
-        
         [7] =
         {
             /* HwunitId:CDD_ADCC Group:0 */
@@ -468,7 +449,6 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
 #endif   
             .soc_num =  (uint8)(1U)
         },
-        
         [8] =
         {
             /* HwunitId:CDD_ADCC Group:1 */
@@ -478,7 +458,7 @@ CONST(struct Cdd_Adc_ConfigTag, CDD_ADC_CONFIG_DATA) Cdd_Adc_Config =
             .extchannelnum = (uint8)(0U),
 #endif   
             .soc_num =  (uint8)(2U)
-        }        
+        }
     },
     .test_input = (Cdd_Adc_InternalTestNodeType)CDD_ADC_TEST_NODE_NO_CONN
 };
